requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
joblib==1.4.0
numpy==1.26.4
pandas==2.2.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.4.0
typing_extensions==4.11.0
tzdata==2024.1
